+++ CHANGELOG.md
@@ -3,9 +3,15 @@
 5.0.7 (unreleased)
 --------------------------
 
+### New Features
+
+- Add `title` and `timestamp` options to the AppleScript’s `write to console` command.
+
+
 ### Improvements
 
 - Provide more Unicode information about the selected character in the document inspector.
+- Don’t hide the console when the application has gone to the background.
 - Optimize file browser’s performance when expanding a folder containing a large number of items.
 - Improve the trimming of whitespaces in the extracted outline labels.
 - Change the syntax highlighting rule to extract strings in the form of “begin-end” string pairs even if one of those is escaped.